apps/api/requirements.txt
# Luminous API Dependencies

# Core
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
pydantic>=2.10.3
pydantic-settings>=2.6.1
python-multipart>=0.0.20
python-dotenv>=1.0.1
httpx>=0.28.1
websockets>=14.1

# AI Services
anthropic>=0.45.1
openai>=1.59.5

# TTS (Text-to-Speech)
# Choose one or more:
# TTS>=0.25.0  # Coqui TTS - Local, high quality, requires setup
# edge-tts>=6.1.12  # Edge-TTS - Free, requires internet
# pyttsx3>=2.90  # Simple, always works offline, limited Thai support

# STT (Speech-to-Text)
# Choose one or more:
openai-whisper>=20240930  # Whisper - Best quality, local
# faster-whisper>=1.1.0  # Faster-Whisper - Faster, less memory
# SpeechRecognition>=3.10.4  # Google Web Speech - Requires internet

# Database (for future use)
sqlalchemy>=2.0.36
asyncpg>=0.30.0
redis>=5.2.1
chromadb>=0.6.3
sentence-transformers>=3.3.1

# Utilities
numpy>=2.2.1
